+++ src/console.ts
@@ -2,8 +2,12 @@ process.env['ENTITIES'] = __dirname + '/**/*.entity.ts';
 
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
+
 import { ProfessionVersionsService } from './professions/profession-versions.service';
 import { ProfessionsSearchService } from './professions/professions-search.service';
+
+import { OrganisationVersionsService } from './organisations/organisation-versions.service';
+import { OrganisationsSearchService } from './organisations/organisations-search.service';
 
 async function bootstrap() {
   const application = await NestFactory.createApplicationContext(AppModule);
@@ -27,6 +31,25 @@ async function bootstrap() {
 
       for (const profession of professions) {
         professionsSearchService.index(profession);
+      }
+
+      break;
+    case 'opensearch:reseed:organisations':
+      console.log('Reseeding organisations...');
+
+      const organisationVersionsService = application.get(
+        OrganisationVersionsService,
+      );
+      const organisationsSearchService = application.get(
+        OrganisationsSearchService,
+      );
+
+      await organisationsSearchService.deleteAll();
+
+      const orgs = await organisationVersionsService.all();
+
+      for (const org of orgs) {
+        organisationsSearchService.index(org);
       }
 
       break;