+++ src/organisations/admin/organisations.controller.spec.ts
@@ -133,11 +133,8 @@ describe('OrganisationsController', () => {
           expect(await controller.index(request)).toEqual(templateParams);
 
           expect(
-            organisationVersionsService.allWithLatestVersion,
-          ).toHaveBeenCalled();
-
-          expect(OrganisationsFilterHelper).toBeCalledWith(organisations);
-          expect(OrganisationsFilterHelper.prototype.filter).toBeCalledWith({
+            organisationVersionsService.searchWithLatestVersion,
+          ).toHaveBeenCalledWith({
             keywords: '',
             nations: [],
             industries: [],
@@ -215,11 +212,8 @@ describe('OrganisationsController', () => {
           ).toEqual(templateParams);
 
           expect(
-            organisationVersionsService.allWithLatestVersion,
-          ).toHaveBeenCalled();
-
-          expect(OrganisationsFilterHelper).toBeCalledWith(organisations);
-          expect(OrganisationsFilterHelper.prototype.filter).toBeCalledWith({
+            organisationVersionsService.searchWithLatestVersion,
+          ).toHaveBeenCalledWith({
             keywords: 'example keywords',
             nations: [nations[2]],
             industries: [industries[1]],
@@ -296,11 +290,8 @@ describe('OrganisationsController', () => {
         expect(await controller.index(request)).toEqual(templateParams);
 
         expect(
-          organisationVersionsService.allWithLatestVersion,
-        ).toHaveBeenCalled();
-
-        expect(OrganisationsFilterHelper).toBeCalledWith(organisations);
-        expect(OrganisationsFilterHelper.prototype.filter).toBeCalledWith({
+          organisationVersionsService.searchWithLatestVersion,
+        ).toHaveBeenCalledWith({
           keywords: '',
           nations: [],
           organisations: [userOrganisation],
@@ -433,10 +424,13 @@ describe('OrganisationsController', () => {
           };
 
           const organisation = organisationFactory.build({ slug: '' });
-          const version = organisationVersionFactory.build({});
+          const version = organisationVersionFactory.build({
+            organisation: organisation,
+          });
 
-          organisationsService.find.mockResolvedValue(organisation);
-          organisationVersionsService.find.mockResolvedValue(version);
+          organisationVersionsService.findByIdWithOrganisation.mockResolvedValue(
+            version,
+          );
 
           const newOrganisation = {
             ...organisation,
@@ -468,7 +462,7 @@ describe('OrganisationsController', () => {
           );
 
           const updatedOrganisation = Organisation.withVersion(
-            organisation,
+            newOrganisation,
             newVersion,
             true,
           );
@@ -514,14 +508,17 @@ describe('OrganisationsController', () => {
             name: 'My awesome organisation',
             slug: 'my-awesome-organisation',
           });
-          const version = organisationVersionFactory.build({});
+          const version = organisationVersionFactory.build({
+            organisation: organisation,
+          });
           const response = createMock<Response>();
           const request = createDefaultMockRequest({
             user: userFactory.build(),
           });
 
-          organisationsService.find.mockResolvedValue(organisation);
-          organisationVersionsService.find.mockResolvedValue(version);
+          organisationVersionsService.findByIdWithOrganisation.mockResolvedValue(
+            version,
+          );
 
           const newVersion = {
             ...version,
@@ -557,7 +554,9 @@ describe('OrganisationsController', () => {
       describe('when the slug is not set', () => {
         beforeEach(async () => {
           organisation = organisationFactory.build({ slug: '' });
-          version = organisationVersionFactory.build();
+          version = organisationVersionFactory.build({
+            organisation: organisation,
+          });
           response = createMock<Response>();
           request = createDefaultMockRequest({
             user: userFactory.build(),
@@ -567,8 +566,9 @@ describe('OrganisationsController', () => {
             confirm: true,
           });
 
-          organisationsService.find.mockResolvedValue(organisation);
-          organisationVersionsService.find.mockResolvedValue(version);
+          organisationVersionsService.findByIdWithOrganisation.mockResolvedValue(
+            version,
+          );
 
           flashMock = flashMessage as jest.Mock;
           flashMock.mockImplementation(() => 'STUB_FLASH_MESSAGE');
@@ -583,17 +583,17 @@ describe('OrganisationsController', () => {
         });
 
         it('should set the slug and confirm the version', async () => {
-          expect(organisationsService.find).toHaveBeenCalledWith('some-uuid');
-          expect(organisationVersionsService.find).toHaveBeenCalledWith(
-            'some-other-uuid',
-          );
+          expect(
+            organisationVersionsService.findByIdWithOrganisation,
+          ).toHaveBeenCalledWith('some-uuid', 'some-other-uuid');
 
           expect(organisationsService.setSlug).toHaveBeenCalledWith(
             organisation,
           );
-          expect(organisationVersionsService.confirm).toHaveBeenCalledWith(
-            version,
-          );
+          expect(organisationVersionsService.confirm).toHaveBeenCalledWith({
+            ...version,
+            organisation: organisation,
+          });
         });
 
         it('should set a flash message and redirect', () => {
@@ -618,7 +618,9 @@ describe('OrganisationsController', () => {
       describe('when the slug is set', () => {
         beforeEach(async () => {
           organisation = organisationFactory.build({ slug: 'some-slug' });
-          version = organisationVersionFactory.build();
+          version = organisationVersionFactory.build({
+            organisation: organisation,
+          });
           response = createMock<Response>();
           request = createDefaultMockRequest({
             user: userFactory.build(),
@@ -628,8 +630,9 @@ describe('OrganisationsController', () => {
             confirm: true,
           });
 
-          organisationsService.find.mockResolvedValue(organisation);
-          organisationVersionsService.find.mockResolvedValue(version);
+          organisationVersionsService.findByIdWithOrganisation.mockResolvedValue(
+            version,
+          );
 
           (escape as jest.Mock).mockImplementation();
 
@@ -649,9 +652,10 @@ describe('OrganisationsController', () => {
           expect(organisationsService.setSlug).not.toHaveBeenCalledWith(
             organisation,
           );
-          expect(organisationVersionsService.confirm).toHaveBeenCalledWith(
-            version,
-          );
+          expect(organisationVersionsService.confirm).toHaveBeenCalledWith({
+            ...version,
+            organisation: organisation,
+          });
         });
 
         it('should set a flash message and redirect', () => {
@@ -687,14 +691,17 @@ describe('OrganisationsController', () => {
     };
 
     const organisation = organisationFactory.build();
-    const version = organisationVersionFactory.build({});
+    const version = organisationVersionFactory.build({
+      organisation: organisation,
+    });
     const response = createMock<Response>();
     const request = createDefaultMockRequest({
       user: userFactory.build(),
     });
 
-    organisationsService.find.mockResolvedValue(organisation);
-    organisationVersionsService.find.mockResolvedValue(version);
+    organisationVersionsService.findByIdWithOrganisation.mockResolvedValue(
+      version,
+    );
 
     await controller.update(
       organisation.id,