+++ src/organisations/admin/organisations.controller.ts
@@ -35,7 +35,6 @@ import { IndexTemplate } from './interfaces/index-template.interface';
 
 import { OrganisationDto } from './dto/organisation.dto';
 import { FilterDto } from './dto/filter.dto';
-import { OrganisationsFilterHelper } from '../helpers/organisations-filter.helper';
 import { IndustriesService } from '../../industries/industries.service';
 import { createFilterInput } from '../../helpers/create-filter-input.helper';
 
@@ -80,8 +79,6 @@ export class OrganisationsController {
       : 'single-organisation';
 
     const allNations = Nation.all();
-    const allOrganisations =
-      await this.organisationVersionsService.allWithLatestVersion();
     const allIndustries = await this.industriesService.all();
 
     const filter = query || new FilterDto();
@@ -96,9 +93,10 @@ export class OrganisationsController {
       filterInput.organisations = [actingUser.organisation];
     }
 
-    const filteredOrganisations = new OrganisationsFilterHelper(
-      allOrganisations,
-    ).filter(filterInput);
+    const filteredOrganisations =
+      await this.organisationVersionsService.searchWithLatestVersion(
+        filterInput,
+      );
 
     const userOrganisation = showAllOrgs
       ? await this.i18nService.translate('app.beis')
@@ -172,13 +170,17 @@ export class OrganisationsController {
     @Res() res: Response,
     @Req() req: RequestWithAppSession,
   ): Promise<void> {
-    const organisation = await this.organisationsService.find(organisationId);
-    const version = await this.organisationVersionsService.find(versionId);
+    const version =
+      await this.organisationVersionsService.findByIdWithOrganisation(
+        organisationId,
+        versionId,
+      );
+    const organisation = version.organisation;
 
     checkCanViewOrganisation(req, organisation);
 
     if (body.confirm) {
-      return this.confirm(res, req, organisation, version);
+      return this.confirm(res, req, version);
     } else {
       if (!isConfirmed(organisation)) {
         organisation.name = body.name;
@@ -205,7 +207,7 @@ export class OrganisationsController {
         this.i18nService,
       );
 
-      return this.showReviewPage(res, organisation, version, {
+      return this.showReviewPage(res, version, {
         ...updatedOrganisation,
         summaryList: await organisationPresenter.summaryList({
           classes: 'govuk-summary-list',
@@ -220,17 +222,17 @@ export class OrganisationsController {
   private async confirm(
     res: Response,
     req: Request,
-    organisation: Organisation,
     version: OrganisationVersion,
   ): Promise<void> {
     let action: string;
 
-    if (!isConfirmed(organisation)) {
+    if (!isConfirmed(version.organisation)) {
       action = 'create';
-      await this.organisationsService.setSlug(organisation);
+      await this.organisationsService.setSlug(version.organisation);
     } else {
       action = 'edit';
     }
+
     await this.organisationVersionsService.confirm(version);
 
     const messageTitle = await this.i18nService.translate(
@@ -239,25 +241,24 @@ export class OrganisationsController {
 
     const messageBody = await this.i18nService.translate(
       `organisations.admin.${action}.confirmation.body`,
-      { args: { name: escape(organisation.name) } },
+      { args: { name: escape(version.organisation.name) } },
     );
 
     req.flash('info', flashMessage(messageTitle, messageBody));
 
     res.redirect(
-      `/admin/organisations/${organisation.id}/versions/${version.id}`,
+      `/admin/organisations/${version.organisation.id}/versions/${version.id}`,
     );
   }
 
   private async showReviewPage(
     res: Response,
-    organisation: Organisation,
     version: OrganisationVersion,
     template: ReviewTemplate,
   ): Promise<void> {
     return res.render('admin/organisations/review', {
       ...template,
-      backLink: `/admin/organisations/${organisation.id}/versions/${version.id}/edit/`,
+      backLink: `/admin/organisations/${version.organisation.id}/versions/${version.id}/edit/`,
     });
   }
 }