+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [unreleased]
+
+### Changed
+
+- Use Opensearch for searching Organisations
+- Backlink on public show pages takes the user back to the search results
+
 ## [release-010] - 2022-03-22
 
 ### Added